feat: rewrite server with h3
