+++ .eslintrc
@@ -4,6 +4,7 @@
     "unicorn/prefer-top-level-await": 0,
     "unicorn/no-process-exit": 0,
     "unicorn/prevent-abbreviations": 0,
+    "unicorn/no-null": 0,
     "no-undef": 0
   }
 }
