+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./ipx";
-export * from "./middleware";
+export * from "./server";
 export * from "./types";
 export * from "./storage/http";
 export * from "./storage/node-fs";
