--- src/middleware.ts
@@ -1,222 +1,0 @@
-import { ServerResponse, IncomingMessage } from "node:http";
-import { negotiate } from "@fastify/accept-negotiator";
-import { decode } from "ufo";
-import getEtag from "etag";
-import xss from "xss";
-import { IPX } from "./ipx";
-import { createError } from "./utils";
-
-const MODIFIER_SEP = /[&,]/g;
-const MODIFIER_VAL_SEP = /[:=_]/;
-
-export interface IPXHRequest {
-  url: string;
-  headers?: Record<string, string>;
-  options?: any;
-}
-
-export interface IPXHResponse {
-  statusCode: number;
-  statusMessage: string;
-  headers: Record<string, string>;
-  body: any;
-  error?: any;
-}
-
-export interface MiddlewareOptions {
-  fallthrough?: boolean;
-}
-
-async function _handleRequest(
-  request: IPXHRequest,
-  ipx: IPX,
-): Promise<IPXHResponse> {
-  const res: IPXHResponse = {
-    statusCode: 200,
-    statusMessage: "",
-    headers: {},
-    body: "",
-  };
-
-  // Parse URL
-  const [modifiersString = "", ...idSegments] = request.url
-    .slice(1 /* leading slash */)
-    .split("/");
-  const id = safeString(decode(idSegments.join("/")));
-
-  // Validate
-  if (!modifiersString) {
-    throw createError("Modifiers are missing", 400, request.url);
-  }
-  if (!id || id === "/") {
-    throw createError("Resource id is missing", 400, request.url);
-  }
-
-  // Contruct modifiers
-  const modifiers: Record<string, string> = Object.create(null);
-
-  // Read modifiers from first segment
-  if (modifiersString !== "_") {
-    for (const p of modifiersString.split(MODIFIER_SEP)) {
-      const [key, ...values] = p.split(MODIFIER_VAL_SEP);
-      modifiers[safeString(key)] = values
-        .map((v) => safeString(decode(v)))
-        .join("_");
-    }
-  }
-
-  // Auto format
-  const mFormat = modifiers.f || modifiers.format;
-  if (mFormat === "auto") {
-    const acceptHeader = request.headers?.accept || "";
-    const autoFormat = autoDetectFormat(
-      acceptHeader,
-      !!(modifiers.a || modifiers.animated),
-    );
-    delete modifiers.f;
-    delete modifiers.format;
-    if (autoFormat) {
-      modifiers.format = autoFormat;
-      res.headers.vary = "Accept";
-    }
-  }
-
-  // Create request
-  const img = ipx(id, modifiers, request.options);
-
-  // Get image meta from source
-  const sourceMeta = await img.getSourceMeta();
-
-  // Caching headers
-  if (sourceMeta.mtime) {
-    if (
-      request.headers?.["if-modified-since"] &&
-      new Date(request.headers["if-modified-since"]) >= sourceMeta.mtime
-    ) {
-      res.statusCode = 304;
-      return res;
-    }
-    res.headers["Last-Modified"] = sourceMeta.mtime.toUTCString();
-  }
-  if (typeof sourceMeta.maxAge === "number") {
-    res.headers[
-      "Cache-Control"
-    ] = `max-age=${+sourceMeta.maxAge}, public, s-maxage=${+sourceMeta.maxAge}`;
-  }
-
-  // Get converted image
-  const { data, format } = await img.process();
-
-  // ETag
-  const etag = getEtag(data);
-  res.headers.ETag = etag;
-  if (etag && request.headers?.["if-none-match"] === etag) {
-    res.statusCode = 304;
-    return res;
-  }
-
-  // Mime
-  if (format) {
-    res.headers["Content-Type"] = `image/${format}`;
-  }
-
-  // Prevent XSS
-  res.headers["Content-Security-Policy"] = "default-src 'none'";
-
-  res.body = data;
-
-  return sanitizeReponse(res);
-}
-
-export function handleRequest(
-  request: IPXHRequest,
-  ipx: IPX,
-): Promise<IPXHResponse> {
-  return _handleRequest(request, ipx).catch((error) => {
-    const statusCode = Number.parseInt(error.statusCode) || 500;
-    // eslint-disable-next-line unicorn/prefer-logical-operator-over-ternary
-    const statusMessage = error.statusMessage
-      ? error.statusMessage
-      : `IPX Error (${statusCode})`;
-    if (process.env.NODE_ENV !== "production" && statusCode === 500) {
-      console.error(error); // eslint-disable-line no-console
-    }
-    return sanitizeReponse({
-      statusCode,
-      statusMessage,
-      body: "IPX Error: " + error,
-      headers: {},
-      error,
-    });
-  });
-}
-
-export function createIPXMiddleware(
-  ipx: IPX,
-  options: Partial<MiddlewareOptions> = {},
-) {
-  return function IPXMiddleware(
-    request: IncomingMessage,
-    res: ServerResponse,
-    next?: (err?: any) => void,
-  ) {
-    return handleRequest(
-      { url: request.url || "/", headers: request.headers as any },
-      ipx,
-    ).then((_res) => {
-      if (options.fallthrough && next && _res.error) {
-        return next(_res.error);
-      }
-      res.statusCode = _res.statusCode;
-      res.statusMessage = _res.statusMessage;
-      for (const name in _res.headers) {
-        res.setHeader(name, _res.headers[name]);
-      }
-      res.end(_res.body);
-    });
-  };
-}
-
-// --- Utils ---
-
-function autoDetectFormat(acceptHeader: string, animated: boolean) {
-  if (animated) {
-    const acceptMime = negotiate(acceptHeader, ["image/webp", "image/gif"]);
-    return acceptMime?.split("/")[1] || "gif";
-  }
-  const acceptMime = negotiate(acceptHeader, [
-    "image/avif",
-    "image/webp",
-    "image/jpeg",
-    "image/png",
-    "image/tiff",
-    "image/heif",
-    "image/gif",
-  ]);
-  return acceptMime?.split("/")[1] || "jpeg";
-}
-
-function sanitizeReponse(res: IPXHResponse) {
-  return <IPXHResponse>{
-    statusCode: res.statusCode || 200,
-    statusMessage: res.statusMessage ? safeString(res.statusMessage) : "OK",
-    headers: safeStringObject(res.headers || {}),
-    body:
-      typeof res.body === "string" ? xss(safeString(res.body)) : res.body || "",
-  };
-}
-
-function safeString(input: string) {
-  return JSON.stringify(input)
-    .replace(/^"|"$/g, "")
-    .replace(/\\+/g, "\\")
-    .replace(/\\"/g, '"');
-}
-
-function safeStringObject(input: Record<string, string>) {
-  const dst: typeof input = {};
-  for (const key in input) {
-    dst[key] = safeString(input[key]);
-  }
-  return dst;
-}
